backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-CORS==3.0.10
Flask-JWT-Extended==4.5.2
SQLAlchemy==2.0.12
mysqlclient==2.1.1
python-dotenv==1.0.0
pycryptodome==3.18.0
marshmallow==3.19.0
celery==5.3.1
redis==4.6.0
# uwsgi==2.0.21
cryptography==41.0.1
werkzeug==2.2.3
alembic==1.11.1
pydantic==1.10.8
pytest==7.3.1
black==23.3.0
flake8==6.0.0
isort==5.12.0
python-dateutil==2.8.2
requests==2.31.0
websockets==11.0.3
pymongo==4.4.0
neo4j==5.9.0 
